+++ core/settlement/engine.go
@@ -46,7 +46,7 @@ type MarketPosition interface {
 //
 //go:generate go run github.com/golang/mock/mockgen -destination mocks/settlement_product_mock.go -package mocks code.vegaprotocol.io/vega/core/settlement Product
 type Product interface {
-	Settle(*num.Uint, *num.Uint, num.Decimal) (*types.FinancialAmount, bool, error)
+	Settle(*num.Uint, *num.Uint, num.Decimal) (*types.FinancialAmount, bool, num.Decimal, error)
 	GetAsset() string
 }
 
@@ -130,17 +130,17 @@ func (e *Engine) Update(positions []events.MarketPosition) {
 }
 
 // Settle run settlement over all the positions.
-func (e *Engine) Settle(t time.Time, settlementData *num.Uint) ([]*types.Transfer, error) {
+func (e *Engine) Settle(t time.Time, settlementData *num.Uint) ([]*types.Transfer, *num.Uint, error) {
 	e.log.Debugf("Settling market, closed at %s", t.Format(time.RFC3339))
-	positions, err := e.settleAll(settlementData)
+	positions, round, err := e.settleAll(settlementData)
 	if err != nil {
 		e.log.Error(
 			"Something went wrong trying to settle positions",
 			logging.Error(err),
 		)
-		return nil, err
+		return nil, nil, err
 	}
-	return positions, nil
+	return positions, round, nil
 }
 
 // AddTrade - this call is required to get the correct MTM settlement values
@@ -403,7 +403,7 @@ func (e *Engine) RemoveDistressed(ctx context.Context, evts []events.Margin) {
 }
 
 // simplified settle call.
-func (e *Engine) settleAll(settlementData *num.Uint) ([]*types.Transfer, error) {
+func (e *Engine) settleAll(settlementData *num.Uint) ([]*types.Transfer, *num.Uint, error) {
 	e.mu.Lock()
 
 	// there should be as many positions as there are parties (obviously)
@@ -418,6 +418,7 @@ func (e *Engine) settleAll(settlementData *num.Uint) ([]*types.Transfer, error) 
 		keys = append(keys, p)
 	}
 	sort.Strings(keys)
+	var delta num.Decimal
 	for _, party := range keys {
 		pos := e.settledPosition[party]
 		// this is possible now, with the Mark to Market stuff, it's possible we've settled any and all positions for a given party
@@ -426,7 +427,7 @@ func (e *Engine) settleAll(settlementData *num.Uint) ([]*types.Transfer, error) 
 		}
 		e.log.Debug("Settling position for party", logging.String("party-id", party))
 		// @TODO - there was something here... the final amount had to be oracle - market or something
-		amt, neg, err := e.product.Settle(e.lastMarkPrice, settlementData.Clone(), num.DecimalFromInt64(pos).Div(e.positionFactor))
+		amt, neg, rem, err := e.product.Settle(e.lastMarkPrice, settlementData.Clone(), num.DecimalFromInt64(pos).Div(e.positionFactor))
 		// for now, product.Settle returns the total value, we need to only settle the delta between a parties current position
 		// and the final price coming from the oracle, so oracle_price - mark_price * volume (check with Tamlyn whether this should be absolute or not)
 		if err != nil {
@@ -436,7 +437,7 @@ func (e *Engine) settleAll(settlementData *num.Uint) ([]*types.Transfer, error) 
 				logging.Error(err),
 			)
 			e.mu.Unlock()
-			return nil, err
+			return nil, nil, err
 		}
 		settlePos := &types.Transfer{
 			Owner:  party,
@@ -451,15 +452,27 @@ func (e *Engine) settleAll(settlementData *num.Uint) ([]*types.Transfer, error) 
 		if neg { // this is a loss transfer
 			settlePos.Type = types.TransferTypeLoss
 			aggregated = append(aggregated, settlePos)
+			// truncated loss amount will not be transferred to the settlement balance
+			// so remove it from the total delta (aka rounding)
+			delta = delta.Sub(rem)
 		} else { // this is a win transfer
 			settlePos.Type = types.TransferTypeWin
 			owed = append(owed, settlePos)
+			// Truncated win transfer won't be withdrawn from the settlement balance
+			// so add it to the total delta (aka rounding)
+			delta = delta.Add(rem)
 		}
+	}
+	// we only care about the int part
+	round := num.UintZero()
+	// if delta > 0, the settlement account will have a non-zero balance at the end
+	if !delta.IsNegative() {
+		round, _ = num.UintFromDecimal(delta)
 	}
 	// append the parties in profit to the end
 	aggregated = append(aggregated, owed...)
 	e.mu.Unlock()
-	return aggregated, nil
+	return aggregated, round, nil
 }
 
 func (e *Engine) getOrCreateCurrentPosition(party string, size int64) (int64, *num.Uint) {